split connection into rendered_connection and connection
